fixes decision when a hash join block finished processing
As it is not guaranteed that a batch will deliver the defined number
of rows (it can be less or more), it’s not possible to make decisions
based only on this.
To solve that we will additionally estimate and track how much batches
of the left relation are required to finish processing 1 block.