Run everything as super user without enterprise enabled.
This makes the `User` non-nullable.

For all practical purposes this shouldn't change the behaviour from a
users point of view.

If enterprise is enabled and host-based authentication is *enabled*, a
user was already mandatory to be present.

If enterprise is enabled and host-based authentication is *disabled*, a
user could connect as `null` user, which behaved the same as if
connected as `crate`. This now changed so that it defaults to `crate`.

If enterprise is disabled, everything ran under a `null` user context.
This also changed to run under a `crate` user. Since there were no
privilege checks without enterprise and functions like `user` are not
available without enterprise, this change is mostly invisible to users,
with the exception of `sys.logs` entries which now default to `crate` as
user as well.

Two main motivations for this change:

 - Being able to avoid null checks for the user

 - Not having to worry if a `null` user means "has no privilege" or "has
 all privileges"